Move token cmd under token new
